Auth

+++ src/main/java/pl/lodz/p/it/eduvirt/service/impl/ResourceGroupServiceImpl.java
@@ -3,16 +3,15 @@ package pl.lodz.p.it.eduvirt.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.ovirt.engine.sdk4.types.Vm;
 import org.ovirt.engine.sdk4.types.VnicProfile;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.lodz.p.it.eduvirt.aspect.logging.LoggerInterceptor;
 import pl.lodz.p.it.eduvirt.dto.nic.NicDto;
 import pl.lodz.p.it.eduvirt.dto.vm.VmDto;
 import pl.lodz.p.it.eduvirt.dto.vm.VmDtoWthEtag;
-import pl.lodz.p.it.eduvirt.entity.Course;
-import pl.lodz.p.it.eduvirt.entity.ResourceGroup;
-import pl.lodz.p.it.eduvirt.entity.ResourceGroupPool;
-import pl.lodz.p.it.eduvirt.entity.VirtualMachine;
+import pl.lodz.p.it.eduvirt.entity.*;
+import pl.lodz.p.it.eduvirt.exceptions.UserNotFoundException;
 import pl.lodz.p.it.eduvirt.exceptions.resource_group.ResourceGroupAlreadyExists;
 import pl.lodz.p.it.eduvirt.exceptions.resource_group.ResourceGroupConflictException;
 import pl.lodz.p.it.eduvirt.exceptions.resource_group.ResourceGroupNotFoundException;
@@ -42,6 +41,7 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     private final PodStatefulRepository podStatefulRepository;
     private final CourseRepository courseRepository;
     private final ETagHelper eTagHelper;
+    private final UserRepository userRepository;
 
     @Override
     @Transactional
@@ -53,6 +53,7 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     public List<VmDto> getVms(UUID id) {
         ResourceGroup resourceGroup = resourceGroupRepository.findById(id)
                 .orElseThrow(() -> new ResourceGroupNotFoundException(id));
+        validateOwnershipOrAdmin(resourceGroup);
         return resourceGroup.getVms()
                 .parallelStream()
                 .map(machine -> {
@@ -70,9 +71,12 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     }
 
     @Override
+    @Transactional
     public VmDtoWthEtag getVm(UUID id) {
         Vm vm = oVirtVmService.findVmById(id.toString());
         VirtualMachine vmEntity = virtualMachineRepository.findById(id).orElseThrow();
+        ResourceGroup resourceGroup = vmEntity.getResourceGroup();
+        validateOwnershipOrAdmin(resourceGroup);
 
         String etag = eTagHelper.generateEtag(vmEntity.getId(), vmEntity.getVersion());
 
@@ -113,7 +117,9 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     @Override
     @Transactional
     public ResourceGroup getResourceGroup(UUID id) {
-        return resourceGroupRepository.findById(id).orElseThrow(() -> new ResourceGroupNotFoundException(id));
+        ResourceGroup resourceGroup = resourceGroupRepository.findById(id).orElseThrow(() -> new ResourceGroupNotFoundException(id));
+        validateOwnershipOrAdmin(resourceGroup);
+        return resourceGroup;
     }
 
     @Override
@@ -128,6 +134,8 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     @Override
     public List<Vm> findAvailableVms(UUID rgId) {
         ResourceGroup resourceGroup = resourceGroupRepository.findById(rgId).orElseThrow(() -> new ResourceGroupNotFoundException(rgId));
+
+        validateOwnership(resourceGroup);
 
         UUID clusterId;
 
@@ -145,6 +153,9 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     @Transactional
     @Override
     public void deleteResourceGroup(UUID id) {
+        ResourceGroup resourceGroup = resourceGroupRepository.findById(id).orElseThrow(() -> new ResourceGroupNotFoundException(id));
+        validateOwnership(resourceGroup);
+
         resourceGroupRepository.deleteById(id);
     }
 
@@ -152,6 +163,8 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
     @Override
     public ResourceGroup updateResourceGroup(UUID id, ResourceGroup resourceGroup, String etag) {
         ResourceGroup existingResourceGroup = resourceGroupRepository.findById(id).orElseThrow(() -> new ResourceGroupNotFoundException(id));
+
+        validateOwnership(existingResourceGroup);
 
         if (!eTagHelper.validateEtag(etag, existingResourceGroup)) {
             throw new ResourceGroupConflictException();
@@ -177,5 +190,49 @@ public class ResourceGroupServiceImpl implements ResourceGroupService {
         existingResourceGroup.setName(resourceGroup.getName());
 
         return resourceGroupRepository.save(existingResourceGroup);
+    }
+
+    @Override
+    @Transactional
+    public void validateResourceGroupOwnership(ResourceGroup resourceGroup) {
+        validateOwnership(resourceGroup);
+    }
+
+    @Transactional
+    @Override
+    public void validateResourceGroupOwnershipOrAdmin(ResourceGroup resourceGroup) {
+        validateOwnershipOrAdmin(resourceGroup);
+    }
+
+    private void validateOwnership(ResourceGroup resourceGroup) {
+        UUID userId = UUID.fromString(SecurityContextHolder.getContext().getAuthentication().getName());
+        boolean isOwner = isOwner(resourceGroup, userId);
+
+        if (!isOwner) {
+            throw new ResourceGroupNotFoundException(resourceGroup.getId());
+        }
+    }
+
+    private void validateOwnershipOrAdmin(ResourceGroup resourceGroup) {
+        UUID userId = UUID.fromString(SecurityContextHolder.getContext().getAuthentication().getName());
+        User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
+
+        boolean isOwner = isOwner(resourceGroup, userId);
+
+        if (!isOwner && !user.getRoles().contains("administrator")) {
+            throw new ResourceGroupNotFoundException(resourceGroup.getId());
+        }
+    }
+
+    private boolean isOwner(ResourceGroup resourceGroup, UUID userId) {
+        boolean isOwner;
+        if (resourceGroup.isStateless()) {
+            ResourceGroupPool pool = resourceGroupPoolRepository.findByResourceGroupsContaining(resourceGroup);
+            isOwner = courseRepository.existsCourseForTeacher(pool.getCourse().getId(), userId);
+        } else {
+            Course course = courseRepository.findByStateFullResourceGroupsContaining(resourceGroup);
+            isOwner = courseRepository.existsCourseForTeacher(course.getId(), userId);
+        }
+        return isOwner;
     }
 }
